requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==4.2.21
django-allauth==65.11.2
django-allauth-ui==1.8.1
django-debug-toolbar==5.2.0
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
PyYAML==6.0.2
requests==2.32.5
slippers==0.6.2
sqlparse==0.5.3
stripe==12.5.1
typeguard==2.13.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
python-dotenv
google-genai
markdown
celery==5.3.6
redis==5.0.1
